pyproject.toml
[project]
name = "tempapp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "zztkm", email = "zztkm@tsurutatakumi.info" }
]
dependencies = [
    "fastapi-slim>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.4.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tempapp"]

[tool.pytest.ini_options]
env_files = "tempapp.env"

[tool.ruff]
exclude = [
    ".eggs",
    "__pypackages__",
    ".mypy_cache",
    ".pytest_cache",
    ".hypothesis",
    ".__pycache__",
]

line-length = 128
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "D100",
    "D104",
    "D400",
    "D415",
    "D203",
    "D213",
    "COM812",
    "ISC001",
    "ANN201",
    "S101",   # assert
]
select = ["ALL"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.isort]
profile = "black"
